Add reverse sql to versioning migration
